RAG/requirements.txt
langchain>=0.3.4
langchain-community>=0.3.4
langchain-text-splitters>=0.3.0
chromadb>=0.5.5
tiktoken>=0.7.0
pypdf>=4.0.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
dashscope>=1.17.0
sentence-transformers>=3.0.1
transformers>=4.44.2
torch>=2.4.0
numpy>=1.26.4
python-dotenv>=1.0.1
rich>=13.7.1
streamlit>=1.38.0